+++ scratch.js
@@ -3,6 +3,8 @@ const readFile = require('./tests/lib/_cachedPage');
 // const wtf = require('./builds/wtf_wikipedia');
 // const wtf = require('./build');
 
+//images in tables...
+
 // wtf.fetch('London', 'en', function(err, doc) {
 //   console.log(doc.sections(0).data);
 // });
@@ -35,18 +37,22 @@ let str = `before
 |}
 |}
 `;
-str = ` {| class="wikitable"
-   |[[File:Worms 01.jpg|199x95px]]
-    |[[File:Worms Wappen 2005-05-27.jpg|199x95px]]
-|<!--col3-->[[File:Liberty-statue-with-manhattan.jpg|199x95px]]
-|<!--col4-->[[File:New-York-Jan2005.jpg|100x95px]]<!--smaller-->
 
 
-  |-
-|<!--col1-->Nibelungen Bridge to Worms
-|Worms and its sister cities
+str = ` {|
+| one
+| two
+| three
+|-
+{|
+| inside! one b
+| inside! two b
+| inside! three b
+|}
 |Statue of Liberty
 |New York City
-|}`;
+|Chicago
+|}
+`;
 
 console.log(wtf(str).tables(0).json());